actualizado

+++ Program.cs
@@ -6,12 +6,11 @@ namespace Promedio
     {
         static void Main(string[] args)
         {
-            
-            int nota1= 55;
-            int nota2= 71;
-            int nota3= 46;
-            int nota4= 87;
-            int promedio= (nota1 + nota2 + nota3 + nota4) / 4;
+            double nota1= 55;
+            double nota2= 71;
+            double nota3= 46;
+            double nota4= 87;
+            double promedio= (nota1 + nota2 + nota3 + nota4) / 4;
 
 
             Console.WriteLine(promedio);
